+++ examples/shared/src/test/scala/zio/examples/test/ExampleSpec.scala
@@ -10,12 +10,6 @@ object ExampleSpec extends DefaultRunnableSpec {
     },
     test("passing test") {
       assert(1)(Assertion.equalTo(1))
-    },
-    test("multiline strings not equal") {
-      assert("foo bar\nbazz buzzard")(Assertion.equalTo("bar foo\nbazard"))
-    },
-    test("multiline string expected to be single line") {
-      assert("Hello,\nWorld!")(Assertion.equalTo("Hello, World!"))
     }
   )
 }
