+++ test-tests/shared/src/test/scala/zio/test/DefaultTestReporterSpec.scala
@@ -3,7 +3,6 @@ package zio.test
 import zio.test.Assertion._
 import zio.test.ReportingTestUtils._
 import zio.test.TestAspect.silent
-import zio.test.diff.Diffing
 
 object DefaultTestReporterSpec extends ZIOBaseSpec {
 
@@ -41,9 +40,6 @@ object DefaultTestReporterSpec extends ZIOBaseSpec {
       },
       testM("correctly reports negated failures") {
         assertM(runLog(test8))(equalTo(test8Expected.mkString + reportStats(0, 0, 1)))
-      },
-      testM("include diff if string equality fails") {
-        assertM(runLog(test9))(equalTo(test9Expected.mkString + reportStats(0, 0, 1), Diffing.noDiffing))
       },
       testM("correctly reports mock failure of invalid argument") {
         assertM(runLog(mock1))(equalTo(mock1Expected.mkString + reportStats(0, 0, 1)))
