+++ test/shared/src/main/scala-2.x/zio/test/AssertionVariants.scala
@@ -17,20 +17,17 @@
 package zio.test
 
 import zio.test.Assertion.Render._
-import zio.test.diff.Diffing
 
 trait AssertionVariants {
 
   /**
    * Makes a new assertion that requires a value equal the specified value.
    */
-  final def equalTo[A, B](expected: A, diffing: Diffing = Diffing.default)(implicit eql: Eql[A, B]): Assertion[B] = {
-    val assertion: Assertion[B] = Assertion.assertion("equalTo")(param(expected)) { actual =>
+  final def equalTo[A, B](expected: A)(implicit eql: Eql[A, B]): Assertion[B] =
+    Assertion.assertion("equalTo")(param(expected)) { actual =>
       (actual, expected) match {
         case (left: Array[_], right: Array[_]) => left.sameElements[Any](right)
         case (left, right)                     => left == right
       }
     }
-    assertion.withDiffing(expected, diffing)
-  }
 }
