--- test/shared/src/main/scala/zio/test/diff/Diffing.scala
@@ -1,18 +1,0 @@
-package zio.test.diff
-
-trait Diffing {
-  def diff[A, B](actual: A, expected: B): Option[DiffResult]
-}
-
-object Diffing {
-  val default = new Diffing {
-    override def diff[A, B](actual: A, expected: B): Option[DiffResult] =
-      (actual, expected) match {
-        case (a: String, e: String) => Some(StringDiffer.default.diff(a, e))
-        case _                      => None
-      }
-  }
-  val noDiffing = new Diffing {
-    override def diff[A, B](actual: A, expected: B): Option[DiffResult] = None
-  }
-}
