--- test/shared/src/main/scala/zio/test/diff/meyers/Equality.scala
@@ -1,8 +1,0 @@
-package zio.test.diff.meyers
-
-trait Equality[T] {
-  def equals(original: T, revised: T): Boolean
-}
-object Equality {
-  def default[T]: Equality[T] = _ == _
-}
